+++ src/main/scala/mesosphere/marathon/core/appinfo/impl/AppInfoBaseData.scala
@@ -162,15 +162,9 @@ class AppInfoBaseData(
   def podStatus(podDef: PodDefinition): Future[PodStatus] = async { // linter:ignore UnnecessaryElseBranch
     val now = clock.now().toOffsetDateTime
     val instances = await(instancesByRunSpecFuture).specInstances(podDef.id)
-    val specByVersion: Map[Timestamp, Option[PodDefinition]] = await(Future.sequence(
-      // TODO(jdef) if repositories ever support a bulk-load interface, use it here
-      instances.map(_.runSpecVersion).distinct.map { version =>
-        groupManager.podVersion(podDef.id, version.toOffsetDateTime).map(version -> _)
-      }
-    )).toMap
     val instanceStatus = instances
       .filter(!_.isScheduled)
-      .flatMap { inst => podInstanceStatus(inst)(specByVersion.apply) }
+      .flatMap { inst => podInstanceStatus(inst) }
     val statusSince = if (instanceStatus.isEmpty) now else instanceStatus.map(_.statusSince).max
     val state = await(podState(podDef.instances, instanceStatus, isPodTerminating(podDef.id)))
 
@@ -223,8 +217,11 @@ class AppInfoBaseData(
     )
   }
 
-  def podInstanceStatus(instance: Instance)(f: Timestamp => Option[PodDefinition]): Option[PodInstanceStatus] = {
-    val maybePodSpec: Option[PodDefinition] = f(instance.runSpecVersion)
+  def podInstanceStatus(instance: Instance): Option[PodInstanceStatus] = {
+    val maybePodSpec: Option[PodDefinition] = instance.runSpec match {
+      case podSpec: PodDefinition => Some(podSpec)
+      case _ => None
+    }
 
     if (maybePodSpec.isEmpty)
       logger.warn(s"failed to generate pod instance status for instance ${instance.instanceId}, " +